requirements.txt
opencv-python==4.12.0.88
numpy==1.26.4
pdf2image==1.17.0
Pillow==11.3.0
google-generativeai==0.8.5
fastapi==0.116.1
paddlepaddle-gpu==2.6.0 -f https://www.paddlepaddle.org.cn/whl/linux/mkl/avx/stable.html
paddleocr==2.7.2
uvicorn